Moved picker functionality to it's own js file Removed inline style from picker
